[_/remove_candle_record] Consolidate OHLCV representation with unified Candle struct
Removes CandleRecord by establishing Candle as the single source of
truth for OHLCV data across engine, strategies, indicators, and
backtests. CandleRow wraps Candle for DB persistence.

Enables indicators to work identically against live and backtest data
without conversion overhead. Type safety improvements: Granularity enum,
consistent volume type.

+++ auro_be/src/engine/live.rs
@@ -165,7 +165,7 @@ pub fn spawn_live_evaluator(mut rx: broadcast::Receiver<StreamMessage>, state: A
                             .bind(closed_candle.high)
                             .bind(closed_candle.low)
                             .bind(closed_candle.close)
-                            .bind(closed_candle.volume as i32)
+                            .bind(closed_candle.volume)
                             .execute(&state.db)
                             .await
                             {
@@ -312,7 +312,7 @@ async fn prefill_buffers(
                 high: *high,
                 low: *low,
                 close: *close,
-                volume: *volume as u32,
+                volume: *volume,
             });
         }
